Clean up test config
