--- eventstore/jpa/src/test/java/org/occurrent/eventstore/jpa/MongoTests.java
@@ -1,10 +1,0 @@
-package org.occurrent.eventstore.jpa;
-
-import org.occurrent.eventstore.jpa.utils.TestHelper;
-
-public class MongoTests extends BaseTest {
-  @Override
-  protected TestHelper testHelper() {
-    return null;
-  }
-}
